Add post processing onnx export test + refactoring
